Module-7-Cloud-Databases/AWS-Kinesis/requirements.txt
boto3
mysql-replication
pymysql